Finish move of MDBook to mdbook-driver
