+++ Cargo.toml
@@ -44,6 +44,7 @@ sha2 = "0.10.9"
 shlex = "1.3.0"
 tempfile = "3.20.0"
 toml = "0.5.11" # Do not update, see https://github.com/rust-lang/mdBook/issues/2037
+topological-sort = "0.2.2"
 
 [package]
 name = "mdbook"
@@ -86,7 +87,6 @@ serde_json.workspace = true
 shlex.workspace = true
 tempfile.workspace = true
 toml.workspace = true
-topological-sort = "0.2.2"
 
 # Watch feature
 notify = { version = "8.0.0", optional = true }
