+++ crates/mdbook-driver/Cargo.toml
@@ -11,12 +11,16 @@ rust-version.workspace = true
 anyhow.workspace = true
 log.workspace = true
 mdbook-core.workspace = true
+mdbook-html.workspace = true
 mdbook-preprocessor.workspace = true
 mdbook-renderer.workspace = true
+mdbook-summary.workspace = true
 regex.workspace = true
 serde_json.workspace = true
 shlex.workspace = true
+tempfile.workspace = true
 toml.workspace = true
+topological-sort.workspace = true
 
 [lints]
 workspace = true
