+++ crates/mdbook-driver/src/lib.rs
@@ -2,3 +2,8 @@
 
 pub mod builtin_preprocessors;
 pub mod builtin_renderers;
+pub mod init;
+mod load;
+mod mdbook;
+
+pub use mdbook::MDBook;
