+++ crates/mdbook-driver/src/mdbook.rs
@@ -1,26 +1,18 @@
-//! The internal representation of a book and infrastructure for loading it from
-//! disk and building it.
-//!
-//! For examples on using `MDBook`, consult the [top-level documentation][1].
-//!
-//! [1]: ../index.html
+//! The high-level interface for loading and rendering books.
 
-mod book;
-mod init;
-
-pub use self::book::load_book;
-pub use self::init::BookBuilder;
+use crate::builtin_preprocessors::{CmdPreprocessor, IndexPreprocessor, LinkPreprocessor};
+use crate::builtin_renderers::{CmdRenderer, MarkdownRenderer};
+use crate::init::BookBuilder;
+use crate::load::{load_book, load_book_from_disk};
 use anyhow::{Context, Error, Result, bail};
 use log::{debug, error, info, log_enabled, trace, warn};
-pub use mdbook_core::book::{Book, BookItem, BookItems, Chapter, SectionNumber};
+use mdbook_core::book::{Book, BookItem, BookItems};
 use mdbook_core::config::{Config, RustEdition};
 use mdbook_core::utils;
-use mdbook_driver::builtin_preprocessors::{CmdPreprocessor, IndexPreprocessor, LinkPreprocessor};
-use mdbook_driver::builtin_renderers::{CmdRenderer, MarkdownRenderer};
 use mdbook_html::HtmlHandlebars;
 use mdbook_preprocessor::{Preprocessor, PreprocessorContext};
 use mdbook_renderer::{RenderContext, Renderer};
-pub use mdbook_summary::{Link, Summary, SummaryItem, parse_summary};
+use mdbook_summary::Summary;
 use std::ffi::OsString;
 use std::io::{IsTerminal, Write};
 use std::path::{Path, PathBuf};
@@ -28,6 +20,9 @@ use std::process::Command;
 use tempfile::Builder as TempFileBuilder;
 use toml::Value;
 use topological_sort::TopologicalSort;
+
+#[cfg(test)]
+mod tests;
 
 /// The object used to manage and build a book.
 pub struct MDBook {
@@ -102,7 +97,7 @@ impl MDBook {
         let root = book_root.into();
 
         let src_dir = root.join(&config.book.src);
-        let book = book::load_book(src_dir, &config.build)?;
+        let book = load_book(src_dir, &config.build)?;
 
         let renderers = determine_renderers(&config);
         let preprocessors = determine_preprocessors(&config)?;
@@ -125,7 +120,7 @@ impl MDBook {
         let root = book_root.into();
 
         let src_dir = root.join(&config.book.src);
-        let book = book::load_book_from_disk(&summary, src_dir)?;
+        let book = load_book_from_disk(&summary, src_dir)?;
 
         let renderers = determine_renderers(&config);
         let preprocessors = determine_preprocessors(&config)?;
@@ -144,8 +139,8 @@ impl MDBook {
     /// `(section: String, bookitem: &BookItem)`
     ///
     /// ```no_run
-    /// # use mdbook::MDBook;
-    /// # use mdbook::book::BookItem;
+    /// # use mdbook_driver::MDBook;
+    /// # use mdbook_core::book::BookItem;
     /// # let book = MDBook::load("mybook").unwrap();
     /// for item in book.iter() {
     ///     match *item {
@@ -617,283 +612,4 @@ fn preprocessor_should_run(
     }
 
     preprocessor.supports_renderer(renderer_name)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::str::FromStr;
-    use toml::value::Table;
-
-    #[test]
-    fn config_defaults_to_html_renderer_if_empty() {
-        let cfg = Config::default();
-
-        // make sure we haven't got anything in the `output` table
-        assert!(cfg.get("output").is_none());
-
-        let got = determine_renderers(&cfg);
-
-        assert_eq!(got.len(), 1);
-        assert_eq!(got[0].name(), "html");
-    }
-
-    #[test]
-    fn add_a_random_renderer_to_the_config() {
-        let mut cfg = Config::default();
-        cfg.set("output.random", Table::new()).unwrap();
-
-        let got = determine_renderers(&cfg);
-
-        assert_eq!(got.len(), 1);
-        assert_eq!(got[0].name(), "random");
-    }
-
-    #[test]
-    fn add_a_random_renderer_with_custom_command_to_the_config() {
-        let mut cfg = Config::default();
-
-        let mut table = Table::new();
-        table.insert("command".to_string(), Value::String("false".to_string()));
-        cfg.set("output.random", table).unwrap();
-
-        let got = determine_renderers(&cfg);
-
-        assert_eq!(got.len(), 1);
-        assert_eq!(got[0].name(), "random");
-    }
-
-    #[test]
-    fn config_defaults_to_link_and_index_preprocessor_if_not_set() {
-        let cfg = Config::default();
-
-        // make sure we haven't got anything in the `preprocessor` table
-        assert!(cfg.get("preprocessor").is_none());
-
-        let got = determine_preprocessors(&cfg);
-
-        assert!(got.is_ok());
-        assert_eq!(got.as_ref().unwrap().len(), 2);
-        assert_eq!(got.as_ref().unwrap()[0].name(), "index");
-        assert_eq!(got.as_ref().unwrap()[1].name(), "links");
-    }
-
-    #[test]
-    fn use_default_preprocessors_works() {
-        let mut cfg = Config::default();
-        cfg.build.use_default_preprocessors = false;
-
-        let got = determine_preprocessors(&cfg).unwrap();
-
-        assert_eq!(got.len(), 0);
-    }
-
-    #[test]
-    fn can_determine_third_party_preprocessors() {
-        let cfg_str = r#"
-        [book]
-        title = "Some Book"
-
-        [preprocessor.random]
-
-        [build]
-        build-dir = "outputs"
-        create-missing = false
-        "#;
-
-        let cfg = Config::from_str(cfg_str).unwrap();
-
-        // make sure the `preprocessor.random` table exists
-        assert!(cfg.get_preprocessor("random").is_some());
-
-        let got = determine_preprocessors(&cfg).unwrap();
-
-        assert!(got.into_iter().any(|p| p.name() == "random"));
-    }
-
-    #[test]
-    fn preprocessors_can_provide_their_own_commands() {
-        let cfg_str = r#"
-        [preprocessor.random]
-        command = "python random.py"
-        "#;
-
-        let cfg = Config::from_str(cfg_str).unwrap();
-
-        // make sure the `preprocessor.random` table exists
-        let random = cfg.get_preprocessor("random").unwrap();
-        let random = get_custom_preprocessor_cmd("random", &Value::Table(random.clone()));
-
-        assert_eq!(random, "python random.py");
-    }
-
-    #[test]
-    fn preprocessor_before_must_be_array() {
-        let cfg_str = r#"
-        [preprocessor.random]
-        before = 0
-        "#;
-
-        let cfg = Config::from_str(cfg_str).unwrap();
-
-        assert!(determine_preprocessors(&cfg).is_err());
-    }
-
-    #[test]
-    fn preprocessor_after_must_be_array() {
-        let cfg_str = r#"
-        [preprocessor.random]
-        after = 0
-        "#;
-
-        let cfg = Config::from_str(cfg_str).unwrap();
-
-        assert!(determine_preprocessors(&cfg).is_err());
-    }
-
-    #[test]
-    fn preprocessor_order_is_honored() {
-        let cfg_str = r#"
-        [preprocessor.random]
-        before = [ "last" ]
-        after = [ "index" ]
-
-        [preprocessor.last]
-        after = [ "links", "index" ]
-        "#;
-
-        let cfg = Config::from_str(cfg_str).unwrap();
-
-        let preprocessors = determine_preprocessors(&cfg).unwrap();
-        let index = |name| {
-            preprocessors
-                .iter()
-                .enumerate()
-                .find(|(_, preprocessor)| preprocessor.name() == name)
-                .unwrap()
-                .0
-        };
-        let assert_before = |before, after| {
-            if index(before) >= index(after) {
-                eprintln!("Preprocessor order:");
-                for preprocessor in &preprocessors {
-                    eprintln!("  {}", preprocessor.name());
-                }
-                panic!("{before} should come before {after}");
-            }
-        };
-
-        assert_before("index", "random");
-        assert_before("index", "last");
-        assert_before("random", "last");
-        assert_before("links", "last");
-    }
-
-    #[test]
-    fn cyclic_dependencies_are_detected() {
-        let cfg_str = r#"
-        [preprocessor.links]
-        before = [ "index" ]
-
-        [preprocessor.index]
-        before = [ "links" ]
-        "#;
-
-        let cfg = Config::from_str(cfg_str).unwrap();
-
-        assert!(determine_preprocessors(&cfg).is_err());
-    }
-
-    #[test]
-    fn dependencies_dont_register_undefined_preprocessors() {
-        let cfg_str = r#"
-        [preprocessor.links]
-        before = [ "random" ]
-        "#;
-
-        let cfg = Config::from_str(cfg_str).unwrap();
-
-        let preprocessors = determine_preprocessors(&cfg).unwrap();
-
-        assert!(
-            !preprocessors
-                .iter()
-                .any(|preprocessor| preprocessor.name() == "random")
-        );
-    }
-
-    #[test]
-    fn dependencies_dont_register_builtin_preprocessors_if_disabled() {
-        let cfg_str = r#"
-        [preprocessor.random]
-        before = [ "links" ]
-
-        [build]
-        use-default-preprocessors = false
-        "#;
-
-        let cfg = Config::from_str(cfg_str).unwrap();
-
-        let preprocessors = determine_preprocessors(&cfg).unwrap();
-
-        assert!(
-            !preprocessors
-                .iter()
-                .any(|preprocessor| preprocessor.name() == "links")
-        );
-    }
-
-    #[test]
-    fn config_respects_preprocessor_selection() {
-        let cfg_str = r#"
-        [preprocessor.links]
-        renderers = ["html"]
-        "#;
-
-        let cfg = Config::from_str(cfg_str).unwrap();
-
-        // double-check that we can access preprocessor.links.renderers[0]
-        let html = cfg
-            .get_preprocessor("links")
-            .and_then(|links| links.get("renderers"))
-            .and_then(Value::as_array)
-            .and_then(|renderers| renderers.get(0))
-            .and_then(Value::as_str)
-            .unwrap();
-        assert_eq!(html, "html");
-        let html_renderer = HtmlHandlebars;
-        let pre = LinkPreprocessor::new();
-
-        let should_run = preprocessor_should_run(&pre, &html_renderer, &cfg);
-        assert!(should_run);
-    }
-
-    struct BoolPreprocessor(bool);
-    impl Preprocessor for BoolPreprocessor {
-        fn name(&self) -> &str {
-            "bool-preprocessor"
-        }
-
-        fn run(&self, _ctx: &PreprocessorContext, _book: Book) -> Result<Book> {
-            unimplemented!()
-        }
-
-        fn supports_renderer(&self, _renderer: &str) -> bool {
-            self.0
-        }
-    }
-
-    #[test]
-    fn preprocessor_should_run_falls_back_to_supports_renderer_method() {
-        let cfg = Config::default();
-        let html = HtmlHandlebars::new();
-
-        let should_be = true;
-        let got = preprocessor_should_run(&BoolPreprocessor(should_be), &html, &cfg);
-        assert_eq!(got, should_be);
-
-        let should_be = false;
-        let got = preprocessor_should_run(&BoolPreprocessor(should_be), &html, &cfg);
-        assert_eq!(got, should_be);
-    }
 }
