+++ src/cmd/clean.rs
@@ -2,7 +2,7 @@ use super::command_prelude::*;
 use crate::get_book_dir;
 use anyhow::Context;
 use anyhow::Result;
-use mdbook::MDBook;
+use mdbook_driver::MDBook;
 use std::mem::take;
 use std::path::PathBuf;
 use std::{fmt, fs};
