+++ src/cmd/serve.rs
@@ -9,8 +9,8 @@ use axum::routing::get;
 use clap::builder::NonEmptyStringValueParser;
 use futures_util::StreamExt;
 use futures_util::sink::SinkExt;
-use mdbook::MDBook;
 use mdbook_core::utils::fs::get_404_output_file;
+use mdbook_driver::MDBook;
 use std::net::{SocketAddr, ToSocketAddrs};
 use std::path::PathBuf;
 use tokio::sync::broadcast;
