+++ src/cmd/test.rs
@@ -3,7 +3,7 @@ use crate::get_book_dir;
 use anyhow::Result;
 use clap::ArgAction;
 use clap::builder::NonEmptyStringValueParser;
-use mdbook::MDBook;
+use mdbook_driver::MDBook;
 use std::path::PathBuf;
 
 // Create clap subcommand arguments
