+++ tests/testsuite/init.rs
@@ -1,7 +1,8 @@
 //! Tests for `mdbook init`.
 
 use crate::prelude::*;
-use mdbook::{Config, MDBook};
+use mdbook_core::config::Config;
+use mdbook_driver::MDBook;
 use std::path::PathBuf;
 
 // Tests "init" with no args.
@@ -18,7 +19,7 @@ All done, no errors...
 
 "#]])
             .expect_stderr(str![[r#"
-[TIMESTAMP] [INFO] (mdbook::book::init): Creating a new book with stub content
+[TIMESTAMP] [INFO] (mdbook_driver::init): Creating a new book with stub content
 
 "#]]);
     })
@@ -84,7 +85,7 @@ All done, no errors...
 
 "#]])
             .expect_stderr(str![[r#"
-[TIMESTAMP] [INFO] (mdbook::book::init): Creating a new book with stub content
+[TIMESTAMP] [INFO] (mdbook_driver::init): Creating a new book with stub content
 
 "#]]);
     })
@@ -116,7 +117,7 @@ All done, no errors...
 
 "#]])
             .expect_stderr(str![[r#"
-[TIMESTAMP] [INFO] (mdbook::book::init): Creating a new book with stub content
+[TIMESTAMP] [INFO] (mdbook_driver::init): Creating a new book with stub content
 
 "#]])
             .args(&["--title", "Example title"]);
