+++ tests/testsuite/redirects.rs
@@ -22,8 +22,8 @@ fn redirects_are_emitted_correctly() {
 fn redirect_removed_with_fragments_only() {
     BookTest::from_dir("redirects/redirect_removed_with_fragments_only").run("build", |cmd| {
         cmd.expect_failure().expect_stderr(str![[r#"
-[TIMESTAMP] [INFO] (mdbook::book): Book building has started
-[TIMESTAMP] [INFO] (mdbook::book): Running the html backend
+[TIMESTAMP] [INFO] (mdbook_driver::mdbook): Book building has started
+[TIMESTAMP] [INFO] (mdbook_driver::mdbook): Running the html backend
 [TIMESTAMP] [ERROR] (mdbook_core::utils): Error: Rendering failed
 [TIMESTAMP] [ERROR] (mdbook_core::utils): [TAB]Caused By: Unable to emit redirects
 [TIMESTAMP] [ERROR] (mdbook_core::utils): [TAB]Caused By: redirect entry for `old-file.html` only has source paths with `#` fragments
@@ -38,8 +38,8 @@ There must be an entry without the `#` fragment to determine the default destina
 fn redirect_existing_page() {
     BookTest::from_dir("redirects/redirect_existing_page").run("build", |cmd| {
         cmd.expect_failure().expect_stderr(str![[r#"
-[TIMESTAMP] [INFO] (mdbook::book): Book building has started
-[TIMESTAMP] [INFO] (mdbook::book): Running the html backend
+[TIMESTAMP] [INFO] (mdbook_driver::mdbook): Book building has started
+[TIMESTAMP] [INFO] (mdbook_driver::mdbook): Running the html backend
 [TIMESTAMP] [ERROR] (mdbook_core::utils): Error: Rendering failed
 [TIMESTAMP] [ERROR] (mdbook_core::utils): [TAB]Caused By: redirect found for existing chapter at `/chapter_1.html`
 Either delete the redirect or remove the chapter.
