+++ tests/testsuite/search.rs
@@ -1,8 +1,7 @@
 //! Tests for search support.
 
 use crate::prelude::*;
-use mdbook::BookItem;
-use mdbook::book::Chapter;
+use mdbook_core::book::{BookItem, Chapter};
 use snapbox::file;
 use std::path::{Path, PathBuf};
 
@@ -135,8 +134,8 @@ fn with_no_source_path() {
 fn chapter_settings_validation_error() {
     BookTest::from_dir("search/chapter_settings_validation_error").run("build", |cmd| {
         cmd.expect_failure().expect_stderr(str![[r#"
-[TIMESTAMP] [INFO] (mdbook::book): Book building has started
-[TIMESTAMP] [INFO] (mdbook::book): Running the html backend
+[TIMESTAMP] [INFO] (mdbook_driver::mdbook): Book building has started
+[TIMESTAMP] [INFO] (mdbook_driver::mdbook): Running the html backend
 [TIMESTAMP] [ERROR] (mdbook_core::utils): Error: Rendering failed
 [TIMESTAMP] [ERROR] (mdbook_core::utils): [TAB]Caused By: [output.html.search.chapter] key `does-not-exist` does not match any chapter paths
