+++ tests/testsuite/test.rs
@@ -7,9 +7,9 @@ use crate::prelude::*;
 fn passing_tests() {
     BookTest::from_dir("test/passing_tests").run("test", |cmd| {
         cmd.expect_stdout(str![[""]]).expect_stderr(str![[r#"
-[TIMESTAMP] [INFO] (mdbook::book): Testing chapter 'Intro': "intro.md"
-[TIMESTAMP] [INFO] (mdbook::book): Testing chapter 'Passing 1': "passing1.md"
-[TIMESTAMP] [INFO] (mdbook::book): Testing chapter 'Passing 2': "passing2.md"
+[TIMESTAMP] [INFO] (mdbook_driver::mdbook): Testing chapter 'Intro': "intro.md"
+[TIMESTAMP] [INFO] (mdbook_driver::mdbook): Testing chapter 'Passing 1': "passing1.md"
+[TIMESTAMP] [INFO] (mdbook_driver::mdbook): Testing chapter 'Passing 2': "passing2.md"
 
 "#]]);
     });
@@ -27,8 +27,8 @@ fn failing_tests() {
             // still includes a little bit of output, so if that is a problem,
             // add more redactions.
             .expect_stderr(str![[r#"
-[TIMESTAMP] [INFO] (mdbook::book): Testing chapter 'Failing Tests': "failing.md"
-[TIMESTAMP] [ERROR] (mdbook::book): rustdoc returned an error:
+[TIMESTAMP] [INFO] (mdbook_driver::mdbook): Testing chapter 'Failing Tests': "failing.md"
+[TIMESTAMP] [ERROR] (mdbook_driver::mdbook): rustdoc returned an error:
 
 --- stdout
 
@@ -38,8 +38,8 @@ test failing.md - Failing_Tests (line 3) ... FAILED
 thread 'main' panicked at failing.md:3:1:
 fail
 ...
-[TIMESTAMP] [INFO] (mdbook::book): Testing chapter 'Failing Include': "failing_include.md"
-[TIMESTAMP] [ERROR] (mdbook::book): rustdoc returned an error:
+[TIMESTAMP] [INFO] (mdbook_driver::mdbook): Testing chapter 'Failing Include': "failing_include.md"
+[TIMESTAMP] [ERROR] (mdbook_driver::mdbook): rustdoc returned an error:
 
 --- stdout
 ...
@@ -62,14 +62,14 @@ fn test_individual_chapter() {
         cmd.args(&["Passing 1"])
             .expect_stdout(str![[""]])
             .expect_stderr(str![[r#"
-[TIMESTAMP] [INFO] (mdbook::book): Testing chapter 'Passing 1': "passing1.md"
+[TIMESTAMP] [INFO] (mdbook_driver::mdbook): Testing chapter 'Passing 1': "passing1.md"
 
 "#]]);
     })
     // Can also be a source path.
     .run("test -c passing2.md", |cmd| {
         cmd.expect_stdout(str![[""]]).expect_stderr(str![[r#"
-[TIMESTAMP] [INFO] (mdbook::book): Testing chapter 'Passing 2': "passing2.md"
+[TIMESTAMP] [INFO] (mdbook_driver::mdbook): Testing chapter 'Passing 2': "passing2.md"
 
 "#]]);
     });
